refactor : 중복코드 함수화 및 리팩토링
