+++ weco-app/pages/mypage/index.tsx
@@ -1,12 +1,12 @@
 import { MainHeader } from 'components/Header';
-import { SessionForm } from 'components/SessionForm';
+import { MyPageForm } from 'components/Form/MyPageForm';
 import { NextPage } from 'next';
 
 const MyPage: NextPage = () => {
   return (
     <>
       <MainHeader />
-      <SessionForm />
+      <MyPageForm />
     </>
   );
 };
